requirements.txt
langchain
langchain-openai
python-dotenv
PyMuPDF
numpy
mistralai
faiss